+++ src/main/java/org/example/shallweeatbackend/filter/CustomLogoutFilter.java
@@ -1,4 +1,4 @@
-package org.example.shallweeatbackend.jwt;
+package org.example.shallweeatbackend.filter;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
@@ -9,6 +9,7 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.example.shallweeatbackend.util.JWTUtil;
 import org.example.shallweeatbackend.repository.RefreshTokenRepository;
 import org.springframework.web.filter.GenericFilterBean;
