+++ src/main/java/org/example/shallweeatbackend/util/JWTUtil.java
@@ -1,4 +1,4 @@
-package org.example.shallweeatbackend.jwt;
+package org.example.shallweeatbackend.util;
 
 import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Value;
